Switch env to settings, embed logic in existing files
