+++ includes/functions.php
@@ -218,4 +218,17 @@
 		return $secret;
 	}
 
+	if ( ! function_exists('env') ) {
+		function env($key, $default = null)
+		{
+			$value = getenv($key);
+  
+			if ($value === false) {
+				return $default;
+			}
+  
+			return $value;
+		}
+	}
+
 ?>
