+++ includes/settings.example.php
